% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshp.R
\name{cshp}
\alias{cshp}
\title{Access the CShapes dataset in R}
\usage{
cshp(date = NA, useGW = TRUE, dependencies = FALSE)
}
\arguments{
\item{date}{The date for which the cshapes polygons should be extracted. This argument must be of type Date and must be in the range 1/1/1886 - end of the dataset. If omitted, the complete dataset is returned.}

\item{useGW}{Boolean argument specifying the system membership coding. TRUE: Gleditsch and Ward (GW, default). FALSE: Correlates of War (COW).}

\item{dependencies}{Boolean argument specifying whether dependent territories must be included. TRUE: Returns polygons for both independent states and dependent units. FALSE: Returns polygons for indepdendent states only (default).}
}
\value{
A \code{sf} dataframe containing the complete CShapes dataset, or the CShapes snapshot for the specified date.
}
\description{
Access the CShapes dataset in R
}
